requirements.txt
streamlit==1.36.0
streamlit-authenticator==0.3.2
pandas==2.1.4
numpy==1.26.4
SQLAlchemy==2.0.31
openai==1.35.13
langchain==0.2.10
langchain-core==0.2.22
langchain-openai==0.1.16
python-dotenv==1.0.1
databricks-sql-connector==3.2.0
httpx==0.27.0